# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voetbalclub Kantine Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface Season {
  id: number;
  name: string;
  startDate: string;
  endDate: string;
  shiftCount: number;
  createdAt: string;
}

export interface CreateSeasonRequest {
  name: string;
  startDate: string;
  endDate: string;
}

export interface ImportResult {
  shiftsCreated: number;
  datesProcessed: number;
  skipped: number;
  message: string;
}

export type VolunteerGroupMembersItem = {
  id: number;
  name: string;
};

export type ShiftSlot = string;

export interface Volunteer {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  /** True als er een wachtwoord is ingesteld voor dit account */
  hasPassword?: boolean;
  /** True als deze vrijwilliger beheerdersrechten heeft */
  isAdmin?: boolean;
  availability: ShiftSlot[];
  groupId?: number | null;
  groupMembers: VolunteerGroupMembersItem[];
  createdAt: string;
}

export interface CreateVolunteerRequest {
  name: string;
  email?: string | null;
  phone?: string | null;
  /** Wachtwoord voor vrijwilliger login (wordt gehasht opgeslagen) */
  password?: string | null;
  /** Geeft de vrijwilliger beheerdersrechten */
  isAdmin?: boolean | null;
  availability?: ShiftSlot[];
  groupMemberIds?: number[] | null;
}

export interface AvailabilitySlot {
  id: number;
  key: string;
  label: string;
  sortOrder: number;
  isActive: boolean;
  /** Starttijd in HH:MM formaat, bijv. 09:00 */
  startTime?: string | null;
  /** Eindtijd in HH:MM formaat, bijv. 13:00 */
  endTime?: string | null;
  createdAt: string;
}

export interface CreateAvailabilitySlotRequest {
  key: string;
  label: string;
  sortOrder?: number | null;
  isActive?: boolean | null;
  startTime?: string | null;
  endTime?: string | null;
}

export interface UpdateAvailabilitySlotRequest {
  label?: string | null;
  sortOrder?: number | null;
  isActive?: boolean | null;
  startTime?: string | null;
  endTime?: string | null;
}

export interface Shift {
  id: number;
  seasonId?: number | null;
  date: string;
  slot: ShiftSlot;
  startTime?: string | null;
  endTime?: string | null;
  capacity: number;
  notes?: string | null;
  createdAt: string;
}

export interface Assignment {
  id: number;
  shiftId: number;
  volunteerId: number;
  volunteer: Volunteer;
  createdAt: string;
}

export type ShiftWithAssignments = Shift & {
  assignments: Assignment[];
};

export interface CreateShiftRequest {
  seasonId?: number | null;
  date: string;
  slot: ShiftSlot;
  startTime?: string | null;
  endTime?: string | null;
  capacity: number;
  notes?: string | null;
}

export interface UpdateShiftRequest {
  date?: string;
  slot?: ShiftSlot;
  startTime?: string | null;
  endTime?: string | null;
  capacity?: number;
  notes?: string | null;
}

export interface AutoScheduleRequest {
  seasonId?: number | null;
}

export interface AutoScheduleResult {
  shiftsScheduled: number;
  assignmentsMade: number;
  message: string;
}

export interface AssignVolunteerRequest {
  volunteerId: number;
}

export interface LoginRequest {
  username?: string | null;
  password?: string | null;
  /** E-mailadres voor vrijwilliger login */
  email?: string | null;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  admin: "admin",
  volunteer: "volunteer",
} as const;

export interface LoginResponse {
  role: LoginResponseRole;
  message: string;
}

export type CurrentUserRole =
  (typeof CurrentUserRole)[keyof typeof CurrentUserRole];

export const CurrentUserRole = {
  admin: "admin",
  volunteer: "volunteer",
} as const;

export interface CurrentUser {
  role: CurrentUserRole;
  username: string;
}

export interface UpdateProfileRequest {
  name?: string;
  email?: string | null;
  phone?: string | null;
  /** Nieuw wachtwoord (leeg = ongewijzigd) */
  password?: string | null;
}

export interface UpdateProfileResponse {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  hasPassword: boolean;
}

export type ImportSeasonScheduleBody = {
  file: Blob;
};

export type ListShiftsParams = {
  startDate?: string;
  endDate?: string;
  seasonId?: number;
};