# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Voetbalclub Kantine Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /availability-slots:
    get:
      operationId: listAvailabilitySlots
      tags: [availability-slots]
      summary: List all availability slot options
      responses:
        "200":
          description: List of availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"
    post:
      operationId: createAvailabilitySlot
      tags: [availability-slots]
      summary: Create a new availability slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilitySlotRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlot"
  /availability-slots/{id}:
    patch:
      operationId: updateAvailabilitySlot
      tags: [availability-slots]
      summary: Update an availability slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilitySlotRequest"
      responses:
        "200":
          description: Updated slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlot"
    delete:
      operationId: deleteAvailabilitySlot
      tags: [availability-slots]
      summary: Delete an availability slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /seasons:
    get:
      operationId: listSeasons
      tags: [seasons]
      summary: List all seasons
      responses:
        "200":
          description: List of seasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Season"
    post:
      operationId: createSeason
      tags: [seasons]
      summary: Create a season
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeasonRequest"
      responses:
        "201":
          description: Season created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
  /seasons/{id}:
    get:
      operationId: getSeason
      tags: [seasons]
      summary: Get a season by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Season found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
        "404":
          description: Not found
    delete:
      operationId: deleteSeason
      tags: [seasons]
      summary: Delete a season
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /seasons/{id}/import:
    post:
      operationId: importSeasonSchedule
      tags: [seasons]
      summary: Import shifts from xlsx file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Shifts imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
...
  /auth/profile:
    patch:
      operationId: updateMyProfile
      tags: [auth]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateProfileResponse"
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    Season:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        shiftCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, startDate, endDate, shiftCount, createdAt]
    CreateSeasonRequest:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
      required: [name, startDate, endDate]
    ImportResult:
      type: object
      properties:
        shiftsCreated:
          type: integer
        datesProcessed:
          type: integer
        skipped:
          type: integer
        message:
          type: string
      required: [shiftsCreated, datesProcessed, skipped, message]
...
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        label:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
        startTime:
          type: string
          nullable: true
          description: "Starttijd in HH:MM formaat, bijv. 09:00"
        endTime:
          type: string
          nullable: true
          description: "Eindtijd in HH:MM formaat, bijv. 13:00"
        createdAt:
          type: string
      required: [id, key, label, sortOrder, isActive, createdAt]
    CreateAvailabilitySlotRequest:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        sortOrder:
          type: integer
          nullable: true
        isActive:
          type: boolean
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
      required: [key, label]
    UpdateAvailabilitySlotRequest:
      type: object
      properties:
        label:
          type: string
          nullable: true
        sortOrder:
          type: integer
          nullable: true
        isActive:
          type: boolean
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
          description: "E-mailadres voor vrijwilliger login"
    LoginResponse:
      type: object
      properties:
        role:
          type: string
          enum: [admin, volunteer]
        message:
          type: string
      required: [role, message]
    CurrentUser:
      type: object
      properties:
        role:
          type: string
          enum: [admin, volunteer]
        username:
          type: string
      required: [role, username]
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
          description: "Nieuw wachtwoord (leeg = ongewijzigd)"
    UpdateProfileResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        hasPassword:
          type: boolean
      required: [id, name, hasPassword]