# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voetbalclub Kantine Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilitySlot {
  id: number;
  key: string;
  label: string;
  sortOrder: number;
  isActive: boolean;
  /** Starttijd in HH:MM formaat, bijv. 09:00 */
  startTime?: string | null;
  /** Eindtijd in HH:MM formaat, bijv. 13:00 */
  endTime?: string | null;
  createdAt: string;
}