# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voetbalclub Kantine Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface ImportResult {
  shiftsCreated: number;
  datesProcessed: number;
  skipped: number;
  message: string;
}